# service/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"BaseResponse_boolean.undefined_": {
				"properties": {
					"metadata": {},
					"message": {
						"type": "string"
					},
					"result": {
						"type": "boolean"
					},
					"code": {
						"$ref": "#/components/schemas/PrefixCode_ObjectDotNotation_BaseCode__"
					}
				},
				"required": [
					"result",
					"code"
				],
				"type": "object"
			},
...
			"BaseResponse_UserSchema-Array._total-number__": {
				"properties": {
					"metadata": {
						"properties": {
							"total": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"total"
						],
						"type": "object"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"items": {
							"$ref": "#/components/schemas/UserSchema"
						},
						"type": "array"
					},
					"code": {
						"$ref": "#/components/schemas/PrefixCode_ObjectDotNotation_BaseCode__"
					}
				},
				"required": [
					"result",
					"code"
				],
				"type": "object"
			},
			"BaseResponse__deleteCnt-number_.undefined_": {
				"properties": {
					"metadata": {},
					"message": {
						"type": "string"
					},
					"result": {
						"properties": {
							"deleteCnt": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"deleteCnt"
						],
						"type": "object"
					},
					"code": {
						"$ref": "#/components/schemas/PrefixCode_ObjectDotNotation_BaseCode__"
					}
				},
				"required": [
					"result",
					"code"
				],
				"type": "object"
			},
			"Express.User": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"BaseResponse_boolean.Express.User_": {
				"properties": {
					"metadata": {
						"$ref": "#/components/schemas/Express.User"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"type": "boolean"
					},
					"code": {
						"$ref": "#/components/schemas/PrefixCode_ObjectDotNotation_BaseCode__"
					}
				},
				"required": [
					"result",
					"code"
				],
				"type": "object"
			},
			"LoginDTO": {
				"properties": {
					"email": {
						"type": "string",
						"format": "email",
						"default": "admin@saige.ai"
					},
					"pwd": {
						"type": "string",
						"default": "admin8282",
						"minLength": 8,
						"maxLength": 50
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "service",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/user/profile": {
			"put": {
				"operationId": "UpdateProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse_boolean.undefined_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"properties": {},
											"additionalProperties": {
												"properties": {
													"message": {
														"type": "string"
													},
													"value": {
														"type": "string"
													}
												},
												"required": [
													"message",
													"value"
												],
												"type": "object"
											},
											"type": "object"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Unhandled",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"name": {
										"type": "string"
									}
								},
								"required": [
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/list": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse_UserSchema-Array._total-number__"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"properties": {},
											"additionalProperties": {
												"properties": {
													"message": {
														"type": "string"
													},
													"value": {
														"type": "string"
													}
												},
												"required": [
													"message",
													"value"
												],
												"type": "object"
											},
											"type": "object"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Unhandled",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "offset",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/user/{idx}": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse__deleteCnt-number_.undefined_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"properties": {},
											"additionalProperties": {
												"properties": {
													"message": {
														"type": "string"
													},
													"value": {
														"type": "string"
													}
												},
												"required": [
													"message",
													"value"
												],
												"type": "object"
											},
											"type": "object"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Unhandled",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idx",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/auth/check": {
			"get": {
				"operationId": "Check",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse_boolean.Express.User_"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"properties": {},
											"additionalProperties": {
												"properties": {
													"message": {
														"type": "string"
													},
													"value": {
														"type": "string"
													}
												},
												"required": [
													"message",
													"value"
												],
												"type": "object"
											},
											"type": "object"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Unhandled",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": []
			}
		},
		"/auth/logout": {
			"get": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse_boolean.undefined_"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"properties": {},
											"additionalProperties": {
												"properties": {
													"message": {
														"type": "string"
													},
													"value": {
														"type": "string"
													}
												},
												"required": [
													"message",
													"value"
												],
												"type": "object"
											},
											"type": "object"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Unhandled",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": []
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse_boolean.undefined_"
								},
								"examples": {
									"admin login sample": {
										"value": {
											"email": "admin@saige.ai",
											"pwd": "admin8282"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"properties": {},
											"additionalProperties": {
												"properties": {
													"message": {
														"type": "string"
													},
													"value": {
														"type": "string"
													}
												},
												"required": [
													"message",
													"value"
												],
												"type": "object"
											},
											"type": "object"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Unhandled",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginDTO"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/service/v1"
		}
	]
}